version_2.0.0/server/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pymongo==4.6.2
python-dotenv==1.0.1
langchain==0.1.12
langchain-community==0.0.27
langchain-groq==0.0.7
sentence-transformers==2.5.1
fitz==0.0.1.dev2
PyMuPDF==1.24.6
opencv-python==4.10.0.84
pytesseract==0.3.10
python-docx==1.1.0
python-magic==0.4.27
langdetect==1.0.9
transformers==4.42.4
spacy==3.7.4
pandas==2.2.0
unstructured==0.12.7
google-api-python-client==2.135.0
google-auth-oauthlib==1.2.1
pyjwt==2.8.0
passlib[bcrypt]==1.7.4 